console: Refactor device view to fix double rendering
