+++ pkg/webui/components/button/button.styl
@@ -25,6 +25,7 @@
   font-size: 1rem
   text-decoration: none
   padding: 0 $cs.s
+  white-space: nowrap
 
   &.primary
     color: white
